DataViews: Refactor the edit function to be based on discrete controls
